+++ src/tests/test_jinja.py
@@ -1,8 +1,8 @@
 import pytest
 
 from datetime import datetime
-from pygw.jinja import Jinja
-from pygw.timetools import to_isotime
+from wxflow.jinja import Jinja
+from wxflow.timetools import to_isotime
 
 current_date = datetime.now()
 j2tmpl = """Hello {{ name }}! {{ greeting }} It is: {{ current_date | to_isotime }}"""
