+++ src/tests/test_timetools.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timedelta
-from pygw.timetools import *
+from wxflow.timetools import *
 
 current_date = datetime.now()
